API-5231: Add support for resolving refs within discriminator mappings
Topic:
https://swagger.io/docs/specification/v3_0/data-models/inheritance-and-polymorphism/#mapping-type-names

This kit completely had no support for it.

The support is now shimmed in wherever testing for a ref, or modifying a ref, occurs. This allows the existing resolution algorithm to proceed more or less as-is.

AI was used to fill in some JSDocs. Treat their accuracy and usefulness accordingly.